[Ref] Silence some 64bit warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@10422 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/ColorConfigDlg.cpp
@@ -144,7 +144,7 @@ BOOL COptionsColors::OnInitDialog()
 	SetDlgItemText(IDC_BUTTON10, FormatFontName(commentFont));
 
 	m_ComboPreset.SetRedraw(FALSE);
-	m_ComboPreset.InitStorage(2 + mpt::size(ColorSchemes), 20 * sizeof(TCHAR));
+	m_ComboPreset.InitStorage(static_cast<int>(2 + mpt::size(ColorSchemes)), 20 * sizeof(TCHAR));
 	m_ComboPreset.AddString(_T("Choose a Colour Scheme..."));
 	m_ComboPreset.AddString(_T("OpenMPT (Default)"));
 	for(const auto &preset : ColorSchemes)
